requirements.txt
please install the following packages at first before running the program!!
---
list:

streamlit
pillow
numpy
plotly
ultralytics